This commit was generated by cvs2svn to compensate for changes in r54, which included commits to RCS files with non-trunk default branches.
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@55 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/arch/mips/mm/pgtable-64.c
@@ -9,7 +9,6 @@
 #include <linux/init.h>
 #include <linux/mm.h>
 #include <asm/pgtable.h>
-#include <asm/pgalloc.h>
 
 void pgd_init(unsigned long page)
 {
@@ -64,7 +63,7 @@ void __init pagetable_init(void)
 
 	memset((void *)kptbl, 0, PAGE_SIZE << PGD_ORDER);
 	memset((void *)kpmdtbl, 0, PAGE_SIZE);
-	set_pgd(swapper_pg_dir, __pgd(kpmdtbl));
+	set_pgd(swapper_pg_dir, __pgd((unsigned long)kpmdtbl));
 
 	/*
 	 * The 64-bit kernel uses a flat pagetable for it's kernel mappings ...
